Extract TestApp to its own module
